test: Use Ubuntu latest (#49)
* test: Run infrastructure on unprivileged ports

Move everything to 8000 + $port, 22 -> 8022, etc. which means it can run
in a rootless container if necessary.

* test: Run on Ubuntu 22.04 (latest)

Ubuntu 22.04 uses cgroups v2 which isn't supported by the version of
systemd used in CentOS 7.x. Therefore:

* Switch the containers to use Rocky 9.x.
* Keep building the images with Docker but run them with Podman as it
  supports running containers with systemd much better than Docker.
  The containers are also run rootless so privileged ports cannot be
  used.
* Enable actions caching for the Docker build steps.
* Make use of some external actions to simplify the workflow.

Podman could be used to build the images however the version installed
lacks a required option for now.